refactor: extract browser command runners
